# api-docs/swagger.yaml
basePath: /cosmos
definitions:
...
  utils.HTTPErrorResult:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: REST API for Cosmos Cloud server management
  title: Cosmos Server API
  version: 0.23.00-unstable001
paths:
...
  /rclone/config/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Remote config with name, type, and parameters
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: Create a new RClone remote
      tags:
      - RClone
  /rclone/config/delete:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Object with name of remote to delete
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: Delete an RClone remote
      tags:
      - RClone
  /rclone/config/dump:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: Dump the full RClone configuration
      tags:
      - RClone
  /rclone/config/listremotes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: List RClone remotes (name and type only, no credentials)
      tags:
      - RClone
  /rclone/config/save:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: Save the RClone configuration
      tags:
      - RClone
  /rclone/config/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Remote update with name and parameters
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: Update an existing RClone remote
      tags:
      - RClone
  /rclone/core/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: Get global RClone transfer stats
      tags:
      - RClone
  /rclone/operations/about:
    post:
      consumes:
      - application/json
      parameters:
      - description: Object with fs field specifying the remote
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: Get storage usage info for a remote (total, used, free)
      tags:
      - RClone
  /rclone/vfs/stats:
    post:
      consumes:
      - application/json
      parameters:
      - description: Object with fs field specifying the remote
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: Get VFS stats for a mounted remote
      tags:
      - RClone
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"